findproject/requirements.txt
asgiref==3.8.1
attrs==26.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
Django==4.2.20
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-extra-fields==3.7.0
drf-spectacular==0.29.0
filetype==1.2.0
flake8==7.2.0
fonttools==4.57.0
fpdf2==2.8.3
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
isort==6.0.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mccabe==0.7.0
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pycparser==2.22
pyflakes==3.3.2
PyJWT==2.9.0
python-dotenv==1.1.0
python3-openid==3.2.0
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.30.0
social-auth-app-django==5.4.3
social-auth-core==4.6.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.4.0
